refactor: 레거시 MVC 제거
