--- app/src/main/java/com/techcourse/adapter/ManualHandlerAdapter.java
@@ -1,24 +1,0 @@
-package com.techcourse.adapter;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.HandlerAdapter;
-import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-import webmvc.org.springframework.web.servlet.view.JspView;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(final Object handler, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
-        final Controller controller = (Controller) handler;
-        final String viewName = controller.execute(request, response);
-
-        return new ModelAndView(new JspView(viewName));
-    }
-}
